Add default feature 'init' to allow minimal builds without all other slog- dependencies
